Extracted an Extractor interface and now using a more proper Decorator pattern as well
