+++ src/main/java/io/arivera/oss/embedded/rabbitmq/extract/CachedExtractor.java
@@ -5,22 +5,21 @@ import io.arivera.oss.embedded.rabbitmq.EmbeddedRabbitMqConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-class CachedExtractor implements Runnable {
+class CachedExtractor extends Extractor.Decorator {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(CachedExtractor.class);
 
-  private final Runnable extractor;
   private final EmbeddedRabbitMqConfig config;
 
-  public CachedExtractor(Runnable extractor, EmbeddedRabbitMqConfig config) {
-    this.extractor = extractor;
+  CachedExtractor(Extractor extractor, EmbeddedRabbitMqConfig config) {
+    super(extractor);
     this.config = config;
   }
 
   @Override
-  public void run() {
+  public void run() throws ExtractionException {
     try {
-      extractor.run();
+      innerExtractor.run();
     } catch (ExtractionException e) {
       if (config.shouldDeleteCachedFileOnErrors()) {
         boolean deleted = config.getDownloadTarget().delete();
